Reduce redundant seed-setting.
The seed was being set for each run of the SSA between a pair of events.
This requires a lot of different seeds, possibly running into the maximum
integer limit in large sets of simulations. It ought to suffice simply to
set the seed once for each simulation, and let this persist through events.